requirements.txt
pydantic ~= 2.4.2
aiohttp ~= 3.8.6
colorama

# Examples
python-dotenv == 1.0.0
sympy ~= 1.12

# Extras
# torch

# Development
# pyyaml ~= 6.0.1
